requirements.txt
fastapi
uvicorn
faiss-cpu
pymupdf
numpy
requests
pydantic
streamlit